i'm back -- began somewhat of rewrite.
I am now taking the modern `ToolipsSession` approach to this. I have some time,  not much, and I am going to set aside a bit of it to work on this package, `ToolipsManager`, and `ToolipsCrawl`.
Crucially, both this package and `ToolipsCrawl` are going to require the next version of `ToolipsSession` to accomplish.
`ToolipsManager` is going to require `Gattino`, and this will also be coming soon. `GattinoInteractive` is also going to heavily depend on this new `ToolipsSession` -- yeah....